requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
fastapi==0.115.12
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-neo4j==0.4.0
langchain-text-splitters==0.3.7
langsmith==0.3.19
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
neo4j==5.28.1
neo4j-graphrag==1.6.1
numpy==2.2.4
orjson==3.10.16
packaging==24.2
propcache==0.3.1
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
pypdf==5.4.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.0.0
types-PyYAML==6.0.12.20250326
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
